feat: move claimsservice tests
